// lib/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
      account: {
        Row: {
          avatar_url: string | null;
          created_at: string;
          email: string;
          full_name: string | null;
          id: string;
          updated_at: string;
        };
        Insert: {
          avatar_url?: string | null;
          created_at?: string;
          email: string;
          full_name?: string | null;
          id?: string;
          updated_at?: string;
        };
        Update: {
          avatar_url?: string | null;
          created_at?: string;
          email?: string;
          full_name?: string | null;
          id?: string;
          updated_at?: string;
        };
        Relationships: [];
      };
      content: {
        Row: {
          account_id: string;
          created_at: string;
          data: Json;
          id: string;
          section: string;
          updated_at: string;
          visibility: Database["public"]["Enums"]["content_visibility"];
        };
        Insert: {
          account_id: string;
          created_at?: string;
          data?: Json;
          id?: string;
          section: string;
          updated_at?: string;
          visibility?: Database["public"]["Enums"]["content_visibility"];
        };
        Update: {
          account_id?: string;
          created_at?: string;
          data?: Json;
          id?: string;
          section?: string;
          updated_at?: string;
          visibility?: Database["public"]["Enums"]["content_visibility"];
        };
        Relationships: [
          {
            foreignKeyName: "content_account_id_fkey";
            columns: ["account_id"];
            isOneToOne: false;
            referencedRelation: "account";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "content_account_id_fkey";
            columns: ["account_id"];
            isOneToOne: false;
            referencedRelation: "public_account_profile";
            referencedColumns: ["id"];
          },
        ];
      };
      setting: {
        Row: {
          account_id: string;
          billing_status: string | null;
          billing_type: string | null;
          created_at: string;
          domain: string | null;
          id: string;
          is_public: boolean;
          preferences: Json;
          updated_at: string;
        };
        Insert: {
          account_id: string;
          billing_status?: string | null;
          billing_type?: string | null;
          created_at?: string;
          domain?: string | null;
          id?: string;
          is_public?: boolean;
          preferences?: Json;
          updated_at?: string;
        };
        Update: {
          account_id?: string;
          billing_status?: string | null;
          billing_type?: string | null;
          created_at?: string;
          domain?: string | null;
          id?: string;
          is_public?: boolean;
          preferences?: Json;
          updated_at?: string;
        };
        Relationships: [
          {
            foreignKeyName: "setting_account_id_fkey";
            columns: ["account_id"];
            isOneToOne: true;
            referencedRelation: "account";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "setting_account_id_fkey";
            columns: ["account_id"];
            isOneToOne: true;
            referencedRelation: "public_account_profile";
            referencedColumns: ["id"];
          },
        ];
      };
    };
    Views: {
      public_account_profile: {
        Row: {
          avatar_url: string | null;
          full_name: string | null;
          id: string | null;
        };
        Relationships: [];
      };
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      content_visibility: "public" | "private";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never;